Return language-specific types from APIs
These changes address this problem: when the user code language is Python, a Python list should be returned from Coco APIs like getEnemies(), instead of a vanilla JavaScript Array.

This adds a per-language hook to convert an object returned from an API to the correct type.  The hook is only implemented for Python so far.

This update requires a 'npm update filbert' to pickup required changes in the Python parser.